Early check for master key length in Secure Cell
Secure Cell forbids empty master keys but if the user happens to use
one then the error reporting may be quite delayed. The error will be
apparent only they the Secure Cell is used for encryption or decryption
while the key has been obtained much earlier.

We're going to perform the check earlier, on SecureCell::with_key()
construction. Now this method returns a Result which will contain an
error if the provided key is empty.

The users now have to check the construction result before using Secure
Cell. Updated tests and documentation examples give an idea of the
interface. Usually it will be enough to simply add "?", but the users
may also opt for "expect()" or "unwrap()" if they know the key is not
empty.