Rename private key structures (#362)
* Rename private key structures

Let's use consistent naming for our key data structure. As outlined
in this pull request comment [1], wording 'secret key' usually refers
to symmetric keys and it is confusing in the context of asymmetric
cryptography. We should use the traditional "private and public keys"
terminology instead.

Update the structure and variable names, command-line arguments and
documentation to consistently use "private keys" everywhere. This is
a breaking change due to structure renaming.

[1]: https://github.com/cossacklabs/themis/pull/358#discussion_r253242588

* Mention breaking changes in changelog

* Fix incorrect memory reservation

Mass-renaming revealed a sleeper bug in memory allocation in unsafe
code. Reserve correct amount of memory for RSA and ECDSA public keys.

Co-Authored-by: ilammy <a.lozovsky@gmail.com>