Early check for master key length in Secure Cell (#365)
* Enforce non-empty keys in KeyBytes

It does not really make sense to have empty key material so let's
enforce this restriction globally. Add length checks to all KeyBytes
constructors and have them return Result which can be an error if
the provided byte slice is empty.

from_vec() constructors are used internally to wrap key bytes that
we receive from key generation functions. These should never be
empty so it's okay to panic if they happen to be empty.

try_from_slice() constructors are intended for users. They can pass
empty slices here, but we do check for that and return an error.
The API already uses result and get_key_kind() already performs
length checks so there are no breaking changes here.

* Early check for master key length in Secure Cell

Secure Cell forbids empty master keys but if the user happens to use
one then the error reporting may be quite delayed. The error will be
apparent only they the Secure Cell is used for encryption or decryption
while the key has been obtained much earlier.

We're going to perform the check earlier, on SecureCell::with_key()
construction. Now this method returns a Result which will contain an
error if the provided key is empty.

The users now have to check the construction result before using Secure
Cell. Updated tests and documentation examples give an idea of the
interface. Usually it will be enough to simply add "?", but the users
may also opt for "expect()" or "unwrap()" if they know the key is not
empty.

* Mention breaking changes in changelog