Link statically against Themis
Dynamic linkage with non-system installation paths is a minefield.
For example, it works incorrectly on macOS if Themis is already
installed to /usr/local/lib (that's because of the "installation path"
thingy in dylibs). On Linux dynamic linkage requires LD_LIBRARY_PATH
to be correctly configured even when building the fuzzing tools.

Use static linkage instead. It turned out to be not that hard. We just
have to add libsoter.a explicitly as well as LDFLAGS from outer Themis
build to pull in the OpenSSL/LibreSSL/BoringSSL dependency.