Fix overflows in Secure Cell (#367)
* Avoid overflow in Secure Cell decryption in sealed mode

Corrupted secure message header or truncated message can trigger an
overflow in message length computation. This integer overflow results
in a segmentation fault later. Handle the overflow and return an error.

We do not have enough information to analyze the header and determine
whether it is corrupted or not. So in case it is corrupted we still
compute a (garbage) message length and then fail decryption if the
provided buffer length is not enough.

Add some tests for Secure Cell to verify authentication context
corruption. Only sealing and token protect mode have authentication
context, there are no tests for context imprint mode.

* Check IV and salt length for symmetric encryption

If the length values are corrupted or incorrect it may trigger
a segmentation fault in cryptographic backend. Both OpenSSL and
BoringSSL are affected.

We should check the length of initialization vector before passing
it to EVP_EncryptInit_ex() or EVP_DecryptInit_ex() because those
functions do not perform any length checks. They expect the buffers
to have suitable lengths (or else they may corrupt the stack).

These functions do not check the length of keys as well, but we
do those checks during key derivation.

We should also check that salt length is zero if there is no salt
because PKCS5_PBKDF2_HMAC() does not check that and may crash due
to NULL dereference.

Add some tests to verify this behavior. Encryption and decryption are
implemented via a common function so it is enough to check only one of
them.