Retain C++03 support
Currenty ThemisPP supports C++03 compilers. Some distributions still use
older compilers that default to C++03. Move semantics are introduced in
C++11 so restrict this feature to C++11 and later. (Obviously, the tests
for moves are no-ops with C++03.)

Copy constructors still need to be disabled though. This is done a bit
differently with C++03 but the effect is the same.