Improve Secure Session ergonomics with STL containers (#369)
* Allow only move for Secure Session objects

Secure Session objects contain pointers to C implementation objects so
they cannot be copied. Implement a proper move constructor for them
and explicitly deny copying.

* Add default constructor for Secure Session objects

This allows to use Secure Session with certain STL containers that
require objects to be default-initializable.

* Throw exception on use-after-move attempts

Check if the Secure Session has been correctly constructed and throw
an exception if the user tries to use a session that has been moved
elsewhere or a default-initialized one.

* Tests for move constructors and assignments

Verify that Secure Session is compatible with some STL containers,
that it is usable after being moved around, and that it throws
exceptions when it should throw them.

* Retain C++03 support

Currenty ThemisPP supports C++03 compilers. Some distributions still use
older compilers that default to C++03. Move semantics are introduced in
C++11 so restrict this feature to C++11 and later. (Obviously, the tests
for moves are no-ops with C++03.)

Copy constructors still need to be disabled though. This is done a bit
differently with C++03 but the effect is the same.