Define non-member global functions as inline
ThemisPP is a header-only library. We should define freestanding
functions as inline in order to trigger the "one-definition rule"
and avoid issues during linkage when multiple instances of the
functions are found in different translation units. This can
happen if multiple translation units include secure_session.hpp.

We don't have to mark class methods in headers as inline, they are
treated as inline automatically.