Forward transport errors from Secure Session callbacks
Now that we are able to store TransportError inside of a themis::Error
we can actually forward the error from the callback to the method call.
With this the user can actually see why the transport layer has failed.

Use the provided SecureSessionContext to temporarily store the error
while the control is still in the C code. Then we extract the error
and return it.

Such implementation does not support concurrent usage of Secure Session
from multiple threads but it's not thread-safe anyway so it is fine to
use this approach.

We can use negative return values to indicate implementation-specific
errors in Secure Session callbacks. The C code returns them as is for
us to inspect. While we're here, stop using bare constant "-1" and give
it a name. Also, handle the overflow with a different error kind.