Add relevant badges and logos (#372)
Brush up and tack some badges onto the Rust package. They will be
visible in many places, like on crates.io, on docs.rs, on GitHub, and
so on. First impression is important. We can add more of these things
later, but they tend to slow down page loads so don't overdo it.

Also, move the rust-themis logo here (it's used by API docs).
Don't link to it via my personal domain, that's a bad practice.