Add MockTransport for testing Secure Session
Rust does not have an established popular mocking framework, and mock
objects are generally frowned upon in tests. However, sometimes they
are useful, and the current practice is to roll out your own mocks.

Introduce a simple implementation of SecureSessionTransport based on
closures. This allows to reimplement and change behavior at run-time.