Switch Soter and Themis makefiles
This has been triggering me for a while, but Soter build commands
belong to soter.mk and Themis build instruction should be located
in themis.mk.

This works now because both files are included into the top-level
makefile, but it is still quite surpising if you are looking to add
some Themis-specific changes to the test makefile.