Do not use uninitialized variables
This is an example of bug prevented by "-Wuninitialized" warning. Here
the "res" variable is not actually initialized during the test failure
handing. If it happens to be non-zero (which is quite likely) then the
failure of a test will not be reported.

secure_session_create() cannot possibly return a themis_status_t and
does not update "res". Just use "false" to always report a failure.