Mark test_rand_with_nist() as unused for CI builds
This function is not actually called by CI due to -DCIRICLE_TEST flag.
(As NIST tests are often failing in Docker environment.) It looks like
dead code to compiler, but I don't want to hide it under an #ifdef so
just mark it as unused instead to simply silence the warning.