Don't build Rust integration tools quietly
Initially I thought that integration tool build should be successful
all the time so I have quieted the "cargo install" command. However,
it turned out that sometimes this build can fail and the error message
is very unhelpful.

Remove the "--quiet" flag and tweak the messages around it. Now the
build will be more verbose and will actually show error messages in
the build logs.