Run C++ tests during CI builds
Well, this is a bit embarrassing, but enable C++ tests in CI builds.
They are there but somehow they were lost on the way.

We'd like to test both C++03 and C++11 flavors of ThemisPP. We will
need to clean the test objects and recompile the tests for both
standards we support. Add a utility Makefile target for that.