Simplify construction of Secure Session
Rename the constructor to just "new()" and make it panic instead of
returning a Result.

There are no user-preventable errors which can be avoided to not cause
a panic. A construction failure usually means either an out-of-memory
condition, or some serious failure in Secure Session invariants or
crypto backend usage by Secure Session. We'd better abort in this case.

As you can see in the tests, we never handle the error and just expect
the construction to always succeed. The user will not get any value out
of error checks like that.