Check for empty client ID in Secure Session
Secure Session is not able to negotiate connection if the client ID
is empty so do an early check and return an error to the user if the
provided client ID is invalid.

Using expect_err() with Result revealed tha SecureSession does not have
a Debug impl. All public types should have one so invent something.