Improve and extend Secure Session API docs (#380)
* Improve and extend Secure Session API docs

Write module-level docs based on description on Wiki. Describe API
flavors of Secure Session and update the method descriptions as well.

* Rename constructor of Secure Session

Rename the constructor to just "new()" to avoid misleading name.

* Reorder Secure Session methods for clarity

Group callback API and buffer-oriented API together in rustdoc output.
This makes navigation a bit easier.

* Rename Secure Session methods for consistency

Use the same naming as other language wrappers do:

- connect_request, wrap, unwrap for buffer-oriented API
- connect, send, receive for callback API

Other wrappers do not have a separate method for negotiation so we have
to invent something here. Keep the naming consistent.

* Use impl Trait in Secure Session methods

Replace explicit templates with impl Trait which makes code a bit
easier to read.

* Check for empty client ID in Secure Session

Secure Session is not able to negotiate connection if the client ID
is empty so do an early check and return an error to the user if the
provided client ID is invalid.

Using expect_err() with Result revealed tha SecureSession does not have
a Debug impl. All public types should have one so invent something.

* Update changelog