Secure Session example code (#381)
Here is an example of using Secure Session to implement a minimal
networking application. Don't do anything fancy: a simple echo server
with fixed keys and basic error handling should be enough.

We need to illustrate the following points:

  - how to initiate connections and exchange data
  - how to use both APIs of SecureSession
  - how to implement SecureSessionTransport correctly

There are also some specific comments on the session termination which
has to be handled on the application layer.