Secure Session: add checks to pubkey callback
It turned out that public key callback did not have enough checks and
could crash Node process if the user returns unexpected value from it.
Add type checks for result: an array is okay, null and undefined are
explicit "not found" values, anything else is also considered "not
found" but with a different error code (the exact value does not matter
at the moment). Also add a length check to avoid buffer overflow when
doing memcpy().