Add Themis status to exceptions
Just like for JavaScript, it may be interesting for the user to check
the status code. Add the status code field to the exception type.

Default status code is "invalid parameter" as that's what most checks
performed by ThemisPP are. Do not use actual default arguments to keep
the ABI stable.