Secure Message: add empty key checks
Just like with JsThemis, Secure Message in ThemisPP has an interface
that may be hard to use correctly. Furthermore, it may be 'successfully'
misused by providing an empty private or public key and then calling
"wrap" or "unwrap" (actually doing "sign" or "verify").

Introduce additional length checks for the keys to ensure that keys are
set when they should be set.

While we're here, use the new status reporting feature of exceptions to
return Themis status codes together with errors.