More paranoid checks for empty data buffers
It is really important to verify that the vectors are non-empty
before taking a pointer to their first element (as in "&data[0]")
because that's undefined behavior if the vector is empty.

The same is true for vector's iterators (data_t::const_iterator). While
it tends to work with current libstdc++ implementations we still should
not dereference invalid iterators as that's undefined behavior as well.