Improve iterator range validation
A likely error with iterators is accidentally misplacing begin and end
iterators in a pair. If we do a check like this then (end, begin) pair
will be considered valid. We should check that the "end" iterator comes
after the "begin" one instead (and that there's at least one element
between them).

This does not protect from all possible errors (e.g., using iterators
from two distinct containers), but it should be good enough for now.