Additional safety checks for Secure Session (#386)
It turns out that Secure Session breaks in a subtle and non-obvious way
if it is created with empty client ID or missing private key. The user
callbacks are also not thoroughly validated and may cause a crash if
the user forgets to fill them in.

Add checks for client ID and private key length as well as NULL checks
for pointers involved with user context and its callbacks.