Remove themis_version() API (#388)
As discussed on GitHub and offline, remove themis_version() function
and all related API for querying Themis and Soter versions at run
time. There is no replacement and this is obviously a breaking change.

This API was not really widely used and it is a bit bothersome to
maintain. It is recommended to not depend on Themis version at run
time. Applications should link against a particular version of Themis
during build or development (whatever is applicable to the language
environment). The version should be selected by the package manager:
either the system one for Themis core, or language-specific for
language bindings.

themis_version() was used tests to print a version banner, that's not
really important. It was also used by Rust binding to verify that FFI
calls can be made, replace it with some other function. And it was also
exported for Ruby programs, drop the export.

Note that the code still contains hardcoded versions in Makefile
as well as in PHP and PHP7 bindings. These are going to stay for
technical reasons.