Keep Rust README minimal (#390)
We have three to four places with assorted documentation on Themis and
it is getting a bit hard to maintain consistency between them. Let's
keep Rust README minimal so that it does not require frequent changes.
Just refer to relevant pieces of documentation from it and avoid any
detailed explanations.